Labs/LAB-3/pyproject.toml
[project]
name = "lab-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "redis-cli>=1.0.1",
]
